Refactor components into Page directories
