+++ src/pages/About/About.js
@@ -1,7 +1,7 @@
-import Education from '../../Education';
+import Education from './Education';
 import NavigationBar from '../../NavigationBar';
-import Skills from '../../Skills';
-import WorkHistory from '../../WorkHistory';
+import Skills from './Skills';
+import WorkHistory from './WorkHistory';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import 'bootstrap/dist/css/bootstrap.min.css';
