+++ src/pages/About/Education.js
